requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
cors==1.0.1
psycopg2-binary==2.9.9
shapely==2.0.3
httpx==0.25.2
pytz==2024.1
redis[asyncio]==5.0.1
slowapi==0.1.9
python-jose[cryptography]==3.3.0